backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
PyPDF2>=3.0.1
python-docx>=0.8.11
numpy>=1.26.0
scikit-learn>=1.3.0
beautifulsoup4>=4.12.2
requests>=2.31.0
fake-useragent>=1.4.0
nltk>=3.8.1
pandas>=2.1.0
setuptools>=69.0.0
wheel>=0.42.0
tensorflow>=2.15.0
mlflow>=2.9.0
kagglehub>=0.1.5 